% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cfg_request.R
\name{cfg-request}
\alias{cfg-request}
\alias{make_cfg_request}
\alias{multi_page_cfg_request}
\title{Core Make Request Function}
\usage{
make_cfg_request(competition, year, page = 1, ...)

multi_page_cfg_request(competition, year, n_pages, ...)
}
\arguments{
\item{competition}{A length-one character vector indicating the competition to query}

\item{year}{A length-one integer vector indicating the year to query}

\item{page}{A length-one integer vectory indicating the page to query. Useful for getting data from the open or the quarterfinals where there are more than 50 competitors.}

\item{...}{Additional parameters passed to the URL as a query string}
}
\description{
This is mostly meant to be an internal function, although in some cases users may want to call it directly
}
