% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossfitgames.R
\docType{package}
\name{crossfitgames-package}
\alias{crossfitgames}
\alias{crossfitgames-package}
\title{crossfitgames: Tools for accessing and shaping data from the Crossfit Games API}
\description{
crossfitgames provides high- and low-level functions to retrieve data from the Crossfit Games API as well as helper functions for extracting and rectangling the retrieved data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ekholme/crossfitgames}
  \item \url{https://ekholme.github.io/crossfitgames/}
  \item Report bugs at \url{https://github.com/ekholme/crossfitgames/issues}
}

}
\author{
\strong{Maintainer}: Eric Ekholm \email{eric.ekholm@gmail.com}

}
\keyword{internal}
