% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_leaderboard.R
\name{extract_final_leaderboard.cfg_leaderboard}
\alias{extract_final_leaderboard.cfg_leaderboard}
\alias{extract_final_leaderboard}
\title{Extract Final Leaderboard}
\usage{
\method{extract_final_leaderboard}{cfg_leaderboard}(obj)

extract_final_leaderboard(x)
}
\arguments{
\item{obj}{A \code{cfg_leaderboard} object}
}
\value{
A 3-column tibble with the rank, name, and total points of competitors
}
\description{
Extract the final leaderboard
}
\examples{
\dontrun{
women_22 <- games_leaderboard(2022, "women")

lb <- extract_final_leaderboard(women_22)
}

}
