% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_workout_results.R
\name{extract_workout_results}
\alias{extract_workout_results}
\title{Extract Workout Results}
\arguments{
\item{obj}{A \code{cfg_leaderboard} object}
}
\value{
A 5-column tibble including
\describe{
\item{athlete}{Athlete name}
\item{workout_num}{Workout number}
\item{workout_place}{Finishing position of the athlete in the workout}
\item{points}{Number of points earned in workout}
\item{score}{Recorded score for workout in time, weight, reps, etc}
}
}
\description{
Extract individual workout results from a \code{cfg_leaderboard} object
}
\examples{
\dontrun{
women_22 <- games_leaderboard(2022, "women")

indiv_results <- extract_workout_results(women_22)
}

}
