% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_leaderboard.R
\name{games_leaderboard}
\alias{games_leaderboard}
\title{Games Leaderboard}
\usage{
games_leaderboard(year, division = c("men", "women", "teams"), ...)
}
\arguments{
\item{year}{The year to request data for, as an integer}

\item{division}{The competition division to request data for. Currently supports "men", "women", and "teams"}

\item{...}{Additional request parameters passed to the API. Not currently used.}
}
\value{
A \code{cfg_leaderboard} object
}
\description{
Retrive a Crossfit Games leaderboard from the Crossfit API. Functions to further process this data are TODO
}
\examples{
\dontrun{
men_22 <- games_leaderboard(2022, division = "men")
}

}
