% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_athlete_id.R
\name{lookup_athlete_id}
\alias{lookup_athlete_id}
\title{Lookup Athlete ID}
\usage{
lookup_athlete_id(name, year, division = c("men", "women", "teams"))
}
\arguments{
\item{name}{Name of athlete to search for}

\item{year}{Year of open competition to search for}

\item{division}{Division to look in. Currently only accepts 'men', 'women', and 'teams'}
}
\value{
A list with information about athletes matching the 'name' parameter.
}
\description{
Utility function to look up potential athlete ID numbers (and other identifying info). This function will be most useful in helping identify the id number of the athlete you're most interested in, which can then (optionally) be passed to the \link{open_leaderboard} function to get Crossfit Open results for a specific athlete
}
\examples{
\dontrun{
#can pass just a last name
res <- lookup_athlete_id("toomey", 2022, "women")

#or you can pass a full name
res2 <- lookup_athlete_id("laura horvath", 2022, "women")

}

}
