% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookups}
\alias{lookups}
\alias{cfg_competitions}
\alias{cfg_divisions}
\alias{cfg_scaled}
\alias{cfg_semis}
\alias{cfg_regions}
\title{Constants to help build queries}
\format{
An object of class \code{list} of length 4.

An object of class \code{list} of length 3.

An object of class \code{list} of length 3.

An object of class \code{list} of length 11.

An object of class \code{list} of length 7.
}
\usage{
cfg_competitions

cfg_divisions

cfg_scaled

cfg_semis

cfg_regions
}
\description{
A subset of the (probably) most useful constants to use to help build queries. Note that this isn't an exhaustive list, but the API isn't well-documented, so it's not straightforward to get everything.
\itemize{
\item \code{cfg_competitions}: values accepted as competition parameters.
\item \code{cfg_divisions}: values accepted as division parameters. Note that this doesn't currently include any teen, masters, or adaptive competitions.
\item \code{cfg_scaled}: values accepted as scaled parameters.
\item \code{cfg_semis}: semi-final competitions. Currently only includes the semifinal competitions from the 2022 season.
\item \code{cfg_regions}: values accepted as region parameters.
}
}
\keyword{datasets}
